add destructuring array into UseFetch type
